[analyzer] NFC: Make reusable unittest mocks reusable.
Put them in a header for other Analyzer unittests to include.

Differential Revision: https://reviews.llvm.org/D60739


git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@358720 91177308-0d34-0410-b5e6-96231b3b80d8
